+++ src/borgitory/api/cloud_sync/cloud_sync.py
@@ -212,162 +212,6 @@ async def get_provider_fields(
     )
 
 
-@router.get("/s3/providers", response_class=HTMLResponse)
-async def get_s3_providers(
-    request: Request,
-    templates: Jinja2Templates = Depends(get_templates),
-) -> HTMLResponse:
-    """Get S3 provider options as HTML"""
-    from borgitory.services.cloud_providers.storage.s3_storage import S3Provider
-    from borgitory.services.cloud_providers.storage.s3_provider_config import (
-        S3ProviderConfig,
-    )
-
-    providers = [
-        {
-            "value": provider.value,
-            "label": S3ProviderConfig.get_provider_label(provider),
-        }
-        for provider in S3Provider
-    ]
-
-    return templates.TemplateResponse(
-        request,
-        "partials/cloud_sync/providers/s3/s3_provider_options.html",
-        {"providers": providers},
-    )
-
-
-@router.get("/s3/regions", response_class=HTMLResponse)
-async def get_s3_regions(
-    request: Request,
-    templates: Jinja2Templates = Depends(get_templates),
-) -> HTMLResponse:
-    """Get S3 regions for a specific provider as HTML"""
-    from borgitory.services.cloud_providers.storage.s3_storage import S3Provider
-    from borgitory.services.cloud_providers.storage.s3_provider_config import (
-        S3ProviderConfig,
-    )
-
-    form_data = await request.form() if request.method == "POST" else {}
-    query_params = dict(request.query_params)
-
-    s3_provider = (
-        form_data.get("provider_config[provider_type]")
-        or query_params.get("s3_provider")
-        or query_params.get("provider_config[provider_type]")
-        or "AWS"
-    )
-    current_value = query_params.get("current_value", "")
-
-    try:
-        provider_enum = S3Provider(s3_provider)
-        regions = S3ProviderConfig.get_regions(provider_enum)
-        default_region = S3ProviderConfig.get_default_region(provider_enum)
-        selected_region = current_value if current_value else default_region
-
-        return templates.TemplateResponse(
-            request,
-            "partials/cloud_sync/providers/s3/s3_region_options.html",
-            {
-                "regions": regions,
-                "selected_region": selected_region,
-                "has_regions": len(regions) > 0,
-            },
-        )
-    except ValueError:
-        return templates.TemplateResponse(
-            request,
-            "partials/cloud_sync/providers/s3/s3_region_options.html",
-            {"regions": [], "selected_region": "us-east-1", "has_regions": False},
-        )
-
-
-@router.get("/s3/storage-classes", response_class=HTMLResponse)
-async def get_s3_storage_classes(
-    request: Request,
-    templates: Jinja2Templates = Depends(get_templates),
-) -> HTMLResponse:
-    """Get storage classes for a specific S3 provider as HTML"""
-    from borgitory.services.cloud_providers.storage.s3_storage import S3Provider
-    from borgitory.services.cloud_providers.storage.s3_provider_config import (
-        S3ProviderConfig,
-    )
-
-    form_data = await request.form() if request.method == "POST" else {}
-    query_params = dict(request.query_params)
-
-    s3_provider = (
-        form_data.get("provider_config[provider_type]")
-        or query_params.get("s3_provider")
-        or query_params.get("provider_config[provider]")
-        or "AWS"
-    )
-    current_value = query_params.get("current_value", "")
-
-    try:
-        provider_enum = S3Provider(s3_provider)
-        storage_classes = S3ProviderConfig.get_storage_classes(provider_enum)
-        default_class = S3ProviderConfig.get_default_storage_class(provider_enum)
-        selected_class = current_value if current_value else default_class
-
-        return templates.TemplateResponse(
-            request,
-            "partials/cloud_sync/providers/s3/s3_storage_class_options.html",
-            {
-                "storage_classes": storage_classes,
-                "selected_class": selected_class,
-            },
-        )
-    except ValueError:
-        return templates.TemplateResponse(
-            request,
-            "partials/cloud_sync/providers/s3/s3_storage_class_options.html",
-        )
-
-
-@router.get("/s3/endpoint-field", response_class=HTMLResponse)
-async def get_s3_endpoint_field(
-    request: Request,
-    templates: Jinja2Templates = Depends(get_templates),
-) -> HTMLResponse:
-    """Get endpoint URL field if required for the selected S3 provider"""
-    from borgitory.services.cloud_providers.storage.s3_storage import S3Provider
-    from borgitory.services.cloud_providers.storage.s3_provider_config import (
-        S3ProviderConfig,
-    )
-
-    form_data = await request.form() if request.method == "POST" else {}
-    query_params = dict(request.query_params)
-
-    s3_provider = (
-        form_data.get("provider_config[provider_type]")
-        or query_params.get("s3_provider")
-        or query_params.get("provider_config[provider_type]")
-        or "AWS"
-    )
-    current_value = query_params.get("current_value", "")
-
-    try:
-        provider_enum = S3Provider(s3_provider)
-        requires_endpoint = S3ProviderConfig.requires_endpoint(provider_enum)
-
-        return templates.TemplateResponse(
-            request,
-            "partials/cloud_sync/providers/s3/s3_endpoint_field.html",
-            {
-                "requires_endpoint": requires_endpoint,
-                "current_value": current_value,
-            },
-        )
-    except ValueError:
-        return templates.TemplateResponse(
-            request,
-            "partials/cloud_sync/providers/s3/s3_endpoint_field.html",
-            {"requires_endpoint": False, "current_value": ""},
-        )
-
-
 @router.post("/", response_class=HTMLResponse)
 async def create_cloud_sync_config(
     request: Request,
