+++ src/borgitory/templates/partials/cloud_sync/providers/s3/s3_provider_options.html
@@ -1,4 +1,4 @@
 {% for provider in providers %}
-<option value="{{ provider.value }}">{{ provider.label }}</option>
+<option value="{{ provider.value }}" {% if provider.value == selected_provider %}selected{% endif %}>{{ provider.label }}</option>
 {% endfor %}
